backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
mariadb>=1.1.8
python-dotenv>=1.0.0
pydantic>=2.5.0
google-generativeai>=0.3.0
httpx>=0.26.0
pandas>=2.1.0
sentence-transformers>=2.3.0
langchain>=0.1.0
langchain-community>=0.0.10
